requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
requests==2.28.1
firebase_admin==6.4.0

google-generativeai

beautifulsoup4==4.11.1
